requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.7
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.6.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
wrapt==1.12.1
